refactor: move DeepNCMDecoder to decoder.py
